Work items are now correctly checked as current or not.

+++ Trousers.Core/Domain/Entities/IVersionable.cs
@@ -2,6 +2,6 @@ namespace Trousers.Core.Domain.Entities
 {
     public interface IVersionable
     {
-        int Revision { get; }
+        long Revision { get; }
     }
 }
